Rename hyp3_auorift.io to prevent shadowing builtin io module
